Minor UI tweaks; Component name refactoring; Fix bug: PlayersScoreBoard now fitting all player names
